# swagger-server/src/main/resources/swagger.yml
swagger: '2.0'
info:
  description: 'Swagger Demo, including swagger-codegen, swagger2markup and asciidoctor.'
  version: v1.1.0
  title: Swagger Demo
  termsOfService: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  contact:
    name: qct
    url: 'http://qct.github.com'
    email: tango.qct@gmail.com
  license:
    name: Apache License
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: '127.0.0.1:8080'
basePath: /
tags:
  - name: book-api
    description: Book Api
  - name: category-api
    description: Category Api
schemes:
  - http
  - https
paths:
  /v1/book/getByCategoryId:
    get:
      tags:
        - book-api
      summary: BookServices
      operationId: getBooksByCategoryUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: query
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Book'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - api_key: []
  /v1/book/update:
    post:
      tags:
        - book-api
      summary: updateBooks
      operationId: updateUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: id
          description: id
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          schema:
            type: boolean
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - api_key: []
  '/v1/book/{id}':
    get:
      tags:
        - book-api
      summary: BookbyId
      operationId: getBookByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Book'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - api_key: []
  /v1/category/:
    get:
      tags:
        - category-api
      summary: getBooks
      operationId: getAllUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Category'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/v1/category/{id}':
    get:
      tags:
        - category-api
      summary: getaBook
      operationId: getUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  Book:
    type: object
    properties:
      author:
        type: string
      categoryId:
        type: integer
        format: int32
      id:
        type: integer
        format: int32
      name:
        type: string
    title: Book
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int32
      name:
        type: string
    title: Category